Fix inheritance issues with VotableThing
